Backend/Libs/requirements.txt
fastapi
uvicorn
pyjwt
pylint
requests
pytest
httpx